refactor: 멀티 모듈 적용을 위한 리뷰 리팩토링
