+++ api/src/main/java/com/programmers/bucketback/domains/review/api/dto/response/ReviewGetByCursorResponse.java
@@ -2,6 +2,7 @@ package com.programmers.bucketback.domains.review.api.dto.response;
 
 import java.util.List;
 
+import com.programmers.bucketback.common.cursor.CursorSummary;
 import com.programmers.bucketback.domains.review.application.dto.ReviewGetByCursorServiceResponse;
 import com.programmers.bucketback.domains.review.model.ReviewCursorSummary;
 
@@ -11,10 +12,12 @@ public record ReviewGetByCursorResponse(
 	List<ReviewCursorSummary> reviews
 ) {
 	public static ReviewGetByCursorResponse from(final ReviewGetByCursorServiceResponse response) {
+		CursorSummary<ReviewCursorSummary> cursorSummary = response.cursorSummary();
+
 		return new ReviewGetByCursorResponse(
 			response.reviewCount(),
-			response.nextCursorId(),
-			response.reviews()
+			cursorSummary.nextCursorId(),
+			cursorSummary.summaries()
 		);
 	}
 }
