+++ api/src/main/java/com/programmers/bucketback/domains/review/application/ReviewService.java
@@ -2,15 +2,18 @@ package com.programmers.bucketback.domains.review.application;
 
 import org.springframework.stereotype.Service;
 
-import com.programmers.bucketback.domains.common.MemberUtils;
-import com.programmers.bucketback.domains.common.vo.CursorPageParameters;
+import com.programmers.bucketback.common.cursor.CursorPageParameters;
+import com.programmers.bucketback.common.cursor.CursorSummary;
 import com.programmers.bucketback.domains.review.application.dto.ReviewGetByCursorServiceResponse;
 import com.programmers.bucketback.domains.review.implementation.ReviewAppender;
 import com.programmers.bucketback.domains.review.implementation.ReviewCursorReader;
 import com.programmers.bucketback.domains.review.implementation.ReviewModifier;
 import com.programmers.bucketback.domains.review.implementation.ReviewRemover;
+import com.programmers.bucketback.domains.review.implementation.ReviewStatistics;
 import com.programmers.bucketback.domains.review.model.ReviewContent;
+import com.programmers.bucketback.domains.review.model.ReviewCursorSummary;
 import com.programmers.bucketback.global.level.PayPoint;
+import com.programmers.bucketback.global.util.MemberUtils;
 
 import lombok.RequiredArgsConstructor;
 
@@ -23,6 +26,7 @@ public class ReviewService {
 	private final ReviewValidator reviewValidator;
 	private final ReviewCursorReader reviewCursorReader;
 	private final ReviewRemover reviewRemover;
+	private final ReviewStatistics reviewStatistics;
 
 	@PayPoint(15)
 	public Long createReview(
@@ -50,7 +54,10 @@ public class ReviewService {
 		final Long itemId,
 		final CursorPageParameters parameters
 	) {
-		return reviewCursorReader.readByCursor(itemId, parameters);
+		Long reviewCount = reviewStatistics.getReviewCount(itemId);
+		CursorSummary<ReviewCursorSummary> cursorSummary = reviewCursorReader.readByCursor(itemId, parameters);
+
+		return new ReviewGetByCursorServiceResponse(reviewCount, cursorSummary);
 	}
 
 	public void deleteReview(
