+++ api/src/main/java/com/programmers/bucketback/domains/review/application/dto/ReviewGetByCursorServiceResponse.java
@@ -1,12 +1,10 @@
 package com.programmers.bucketback.domains.review.application.dto;
 
-import java.util.List;
-
+import com.programmers.bucketback.common.cursor.CursorSummary;
 import com.programmers.bucketback.domains.review.model.ReviewCursorSummary;
 
 public record ReviewGetByCursorServiceResponse(
 	Long reviewCount,
-	String nextCursorId,
-	List<ReviewCursorSummary> reviews
+	CursorSummary<ReviewCursorSummary> cursorSummary
 ) {
 }
